requirements.txt
python-dotenv
anthropic
openai
tenacity
tqdm
pandas
chardet
